Adds code to handle more CallExpr cases
Previously code would dump many "odd, we found a fdecl without a callee"
messages to the console as it found call expressions that were not the typical
function(arg,umen,ts) format. The code is now able to identify what other types
of call expressions it has, but it does not do anything useful with them at the
moment. Most notably, the Constructor, Destructor, and Function Pointer calls
should be investigated to yeild better analysis.