Adds basic sfpf binary
sfpf is the Static Function Property Finder and it utilizes a modified deduction
engine, which propigates information from library functions through defined
functions. This is the reverse direction compared to the deductions used in
sfpv, thus the ReverseDedutions class. The ideal use of sfpf would be to quickly
obtain whitelists and blacklists from a library.

- Adds basic sfpf binary
- Moves main.cpp to sfpv.cpp
- Adds simple internal blacklist in sfpf