Adds basic support for function pointer tracking
Function pointers are now analyzed, making this project track direct, indirect,
and virtual calls. In order to do this some of the hacks to allow the lookup of
particular callees were undone and the underlying system was made to be a bit
more general. This amounted to useing a large number of pointers to Callee
objects rather than accessing them based upon their fully qualified names.
Unfortuantly this does introduce the possiblilty of unexpected segfaults, though
the test suite appears to work properly and deallocation is not handled in the
current revision (aka non of the objects are deleted after being newed).

- Function pointers are now recorded and the possible calls that they contain
  are stored
- sfpf was broken in this rework, though it should not be difficult to fix in
  another commit
- Test added to check that function pointers are added to the callgraph
  correctly