Adds tracking of function pointers through calls
Now if a function pointer is passed into a function, it is tracked.
The current implementation is rather simplistic and it will fail when names
conflict, but the basic behavior is now there. In order to test this
functionality function pointers now have their attributes checked (ie. they can
be marked realtime or non-realtime). This support should be able to
pessimistically track functions through callbacks.

- Adds tracking of function pointers through function calls
- Adds annotation checking for function pointer
- Adds test for annotated function pointers