Fixes compile issues with clang 3.2
Apparently the old API is no longer valid and some memory management has changed
in the bowels of libclang.
Thus to stop the fires from some refcounting gone wrong in TargetOptions or an
associated class cleanup is simply not done.
All the memory is reclaimed by the OS, but is still quite annoying that this
commit had to be made.