Misc Updates For LLVM 3.5
I really wish that there was a more sane way to get the clang linker flags
setup, but they seem to change with every release and due to single pass
linkers, the ordering is a PITA to get right.
None the less, it should work with llvm 3.5 as compiled from scratch.